requirements.txt
fire
loguru
fire
matplotlib
nats-py
numpy
pre-commit
psycopg2-binary
pydantic
pydantic-settings
pytest
pytest-cov
pytest-postgresql
setuptools
wheel
